api's for admin dash are created

+++ src/controllers/auth.controller.js
@@ -34,11 +34,11 @@ export const registerUser = async (req, res) => {
         const token = generateToken(newUser._id, newUser.role)
         // Set the token in a cookie
         res.cookie('token', token,
-            {
-                httpOnly: true,
-                secure: true,
-                sameSite: 'none'
-            }
+            // {
+            //     httpOnly: true,
+            //     secure: true,
+            //     sameSite: 'none'
+            // }
         );
         res.status(201).json({
             message: "User created successfully", user: {
@@ -74,11 +74,11 @@ export const loginUser = async (req, res) => {
         const token = generateToken(user._id, user.role)
         // Set the token in a cookie
         res.cookie('token', token,
-            {
-                httpOnly: true,
-                secure: true,
-                sameSite: 'none'
-            }
+            // {
+            //     httpOnly: true,
+            //     secure: true,
+            //     sameSite: 'none'
+            // }
         );
         res.status(200).json({
             message: "Login successful", user: {
@@ -93,12 +93,51 @@ export const loginUser = async (req, res) => {
     }
 }
 /**
+ * @desc Admin register user
+ * @route POST /api/users/admin-register
+ * @access Private (Admin only)
+ */
+export const addUser = async (req, res) => {
+    const { username, email, password, role = "user" } = req.body
+    try {
+        if (!username || !email || !password) {
+            return res.status(400).json({ message: "All fields are required" })
+        }
+        if (password.length < 6) {
+            return res.status(400).json({ message: "Password must be at least 6 characters long" })
+        }
+        const existingUser = await userSchema.findOne({
+            $or: [
+                { username: username },
+                { email: email }
+            ]
+        })
+        if (existingUser) {
+            return res.status(400).json({ message: "User already exists" })
+        }
+        const hashedPassword = await bcrypt.hash(password, 10)
+        const newUser = await userSchema.create({
+            username,
+            email,
+            role,
+            password: hashedPassword
+        })
+
+        res.status(201).json({
+            message: "User created successfully"
+        })
+    } catch (error) {
+        res.status(500).json({ message: "Internal server error" })
+    }
+}
+/**
  * @desc Admin login user
  * @route POST /api/users/admin-login
  * @access Private (Admin only)
  */
 export const adminLoginUser = async (req, res) => {
     const { username, password } = req.body
+
     try {
         if (!username || !password) {
             return res.status(400).json({ message: "All fields are required" })
@@ -113,6 +152,21 @@ export const adminLoginUser = async (req, res) => {
         const isPasswordValid = await bcrypt.compare(password, user.password)
         if (!isPasswordValid) {
             return res.status(401).json({ message: "Invalid email or password" })
+        }
+        const token = generateToken(user._id, user.role)
+        // Set the token in a cookie
+        res.cookie('token', token,
+            {
+                httpOnly: true,
+                secure: true,
+                sameSite: 'none'
+            }
+        );
+
+        const decoded = req.user
+
+        if (decoded.role !== "admin") {
+            return res.status(403).json({ message: "Forbidden: Admins only" })
         }
 
         res.status(200).json({
@@ -158,6 +212,51 @@ export const getUserProfile = async (req, res) => {
 
 }
 /**
+ * @desc Update user profile
+ * @route Patch /api/users/profile
+ * @access Private (user only)
+ */
+export const updateUserProfile = async (req, res) => {
+    const { id } = req.params
+    const { username, email, password, role = "user" } = req.body
+    try {
+        const user = req.user
+        if (!user) {
+            return res.status(404).json({ message: "User not found" })
+        }
+        if (user.role !== "admin") {
+            return res.status(403).json({ message: "Forbidden: Only admins can change role to admin" })
+        }
+        if (password && password.length < 6) {
+            return res.status(400).json({ message: "Password must be at least 6 characters long" })
+        }
+        const hashedPassword = await bcrypt.hash(password, 10)
+        const updatedUser = await userSchema.findByIdAndUpdate(id,
+            {
+                username,
+                email,
+                role,
+                password: hashedPassword
+            }, {
+            returnDocument: 'after'
+        })
+        if (!updatedUser) {
+            return res.status(404).json({ message: "User not found" })
+        }
+        res.status(200).json({
+            user: {
+                id: updatedUser._id,
+                username: updatedUser.username,
+                email: updatedUser.email,
+                role: updatedUser.role
+            }
+        })
+
+    } catch (error) {
+        res.status(500).json({ message: "Internal server error" })
+    }
+}
+/**
  * @desc Logout user
  * @route POST /api/users/logout
  * @access Private (user only)
@@ -191,6 +290,13 @@ export const deleteUser = async (req, res) => {
         if (user.role !== "admin") {
             return res.status(403).json({ message: "Forbidden: Admins only" })
         }
+        const userToDelete = await userSchema.findById(id);
+        if (!userToDelete) {
+            return res.status(404).json({ message: "User not found" })
+        }
+        if (userToDelete.role === "admin") {
+            return res.status(403).json({ message: "Forbidden: Cannot delete admin user" })
+        }
         const deletedUser = await userSchema.findByIdAndDelete(id);
         if (!deletedUser) {
             return res.status(404).json({
@@ -198,7 +304,8 @@ export const deleteUser = async (req, res) => {
             });
         }
         res.status(200).json({
-            message: "User deleted successfully"
+            message: "User deleted successfully",
+            role: deletedUser.role
         });
     } catch (error) {
         console.error("Delete user error:", error);
@@ -219,7 +326,7 @@ export const getAllUsers = async (req, res) => {
             return res.status(403).json({ message: "Forbidden: Admins only" })
         }
         const users = await userSchema.find() // Exclude password field
-        res.status(200).json({ users })
+        res.status(200).json(users)
     } catch (error) {
         res.status(500).json({ message: "Internal server error" })
     }
